requirements.txt
python-telegram-bot==20.3
requests
SQLAlchemy==2.0.36
psycopg[binary]
python-dotenv==1.0.1
